auditor/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
psycopg2-binary==2.9.9
pydantic==2.8.0
pydantic-settings==2.3.0
anthropic==0.28.1
semgrep==1.75.0
radon==6.0.1
python-multipart==0.0.9
